+++ src/dialog/error.ts
@@ -1,0 +1,6 @@
+export class DialogValidationError extends Error {
+  public constructor(message?: string) {
+    super(message)
+    this.name = 'DialogValidationError'
+  }
+}