+++ src/dialog/index.ts
@@ -1,0 +1,19 @@
+// Dialog container
+export { Dialog } from './Dialog'
+
+// Dialog form elements
+export { Input } from './Input'
+export { Textarea } from './Textarea'
+export {
+  Select,
+  SelectFragment,
+  Option,
+  Optgroup,
+  ExternalSelect,
+  UsersSelect,
+  ConversationsSelect,
+  ChannelsSelect,
+} from './Select'
+
+// Error class
+export { DialogValidationError } from './error'